@@ -2,39 +2,38 @@
 
 namespace PanKrok\ShoperAppstoreBundle\Entity;
 
-use Doctrine\ORM\Mapping as ORM;
 use PanKrok\ShoperAppstoreBundle\Repository\SubscriptionsRepository;
+use Doctrine\ORM\Mapping as ORM;
 
-/**
- * @ORM\Entity(repositoryClass=SubscriptionsRepository::class)
- */
+#[ORM\Entity(repositoryClass: SubscriptionsRepository::class)]
 class Subscriptions
 {
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     */
-    private $id;
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column]
+    private ?int $id = null;
 
-    /**
-     * @ORM\ManyToOne(targetEntity=Shops::class, inversedBy="subscriptions")
-     */
-    private $shop;
+    #[ORM\Column(nullable: true)]
+    private ?\DateTimeImmutable $expires_at = null;
 
-    /**
-     * @ORM\Column(type="datetime", nullable=true)
-     */
-    private $expires_at;
-    
-    public function __toString()
-    {
-        return $this->id;
-    }
+    #[ORM\ManyToOne(inversedBy: 'subscriptions')]
+    private ?Shops $shop = null;
 
     public function getId(): ?int
     {
         return $this->id;
+    }
+
+    public function getExpiresAt(): ?\DateTimeImmutable
+    {
+        return $this->expires_at;
+    }
+
+    public function setExpiresAt(?\DateTimeImmutable $expires_at): static
+    {
+        $this->expires_at = $expires_at;
+
+        return $this;
     }
 
     public function getShop(): ?Shops
@@ -42,21 +41,9 @@ class Subscriptions
         return $this->shop;
     }
 
-    public function setShop(?Shops $shop): self
+    public function setShop(?Shops $shop): static
     {
         $this->shop = $shop;
-
-        return $this;
-    }
-
-    public function getExpiresAt(): ?\DateTimeInterface
-    {
-        return $this->expires_at;
-    }
-
-    public function setExpiresAt(?\DateTimeInterface $expires_at): self
-    {
-        $this->expires_at = $expires_at;
 
         return $this;
     }
