@@ -2,15 +2,12 @@
 
 namespace PanKrok\ShoperAppstoreBundle\Repository;
 
+use PanKrok\ShoperAppstoreBundle\Entity\Billings;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
-use PanKrok\ShoperAppstoreBundle\Entity\Billings;
 
 /**
- * @method Billings|null find($id, $lockMode = null, $lockVersion = null)
- * @method Billings|null findOneBy(array $criteria, array $orderBy = null)
- * @method Billings[]    findAll()
- * @method Billings[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ * @extends ServiceEntityRepository<Billings>
  */
 class BillingsRepository extends ServiceEntityRepository
 {
@@ -19,32 +16,28 @@ class BillingsRepository extends ServiceEntityRepository
         parent::__construct($registry, Billings::class);
     }
 
-    // /**
-    //  * @return Billings[] Returns an array of Billings objects
-    //  */
-    /*
-    public function findByExampleField($value)
-    {
-        return $this->createQueryBuilder('b')
-            ->andWhere('b.exampleField = :val')
-            ->setParameter('val', $value)
-            ->orderBy('b.id', 'ASC')
-            ->setMaxResults(10)
-            ->getQuery()
-            ->getResult()
-        ;
-    }
-    */
+    //    /**
+    //     * @return Billings[] Returns an array of Billings objects
+    //     */
+    //    public function findByExampleField($value): array
+    //    {
+    //        return $this->createQueryBuilder('b')
+    //            ->andWhere('b.exampleField = :val')
+    //            ->setParameter('val', $value)
+    //            ->orderBy('b.id', 'ASC')
+    //            ->setMaxResults(10)
+    //            ->getQuery()
+    //            ->getResult()
+    //        ;
+    //    }
 
-    /*
-    public function findOneBySomeField($value): ?Billings
-    {
-        return $this->createQueryBuilder('b')
-            ->andWhere('b.exampleField = :val')
-            ->setParameter('val', $value)
-            ->getQuery()
-            ->getOneOrNullResult()
-        ;
-    }
-    */
+    //    public function findOneBySomeField($value): ?Billings
+    //    {
+    //        return $this->createQueryBuilder('b')
+    //            ->andWhere('b.exampleField = :val')
+    //            ->setParameter('val', $value)
+    //            ->getQuery()
+    //            ->getOneOrNullResult()
+    //        ;
+    //    }
 }
